requirements.txt
torch~=1.9.1+cu111
torchvision~=0.10.1+cu111
numpy~=1.22.3
albumentations~=1.1.0
torchsummary~=1.5.1
python-dotenv~=0.20.0